Store 간단한 save메소드 + validation + 반경 내 가게들 찾기 + 사업자 등록번호로 가게 찾기:ver 2
