+++ api/throwapi/src/main/java/sosteam/throwapi/domain/store/entity/Store.java
@@ -3,13 +3,16 @@ package sosteam.throwapi.domain.store.entity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
+import lombok.AccessLevel;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 import sosteam.throwapi.domain.user.entity.User;
 import sosteam.throwapi.global.entity.PrimaryKeyEntity;
 
 @Getter
 @Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Store extends PrimaryKeyEntity {
 
     @NotNull
@@ -17,8 +20,6 @@ public class Store extends PrimaryKeyEntity {
 
     @NotNull
     private String companyRegistrationNumber;
-
-    public Store() {};
 
     public Store(String name, String companyRegistrationNumber) {
         this.name = name;
