+++ api/throwapi/src/main/java/sosteam/throwapi/global/exception/exception/CommonException.java
@@ -1,11 +1,14 @@
 package sosteam.throwapi.global.exception.exception;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.springframework.http.HttpStatus;
 
 @Data
+@AllArgsConstructor
 public class CommonException extends RuntimeException{
     private final String code;
-    private final String message;
+    private String message;
     private final HttpStatus status;
 }
