+++ api/throwapi/src/main/java/sosteam/throwapi/global/exception/exception/NoContentException.java
@@ -1,10 +1,10 @@
-package sosteam.throwapi.domain.store.exception;
+package sosteam.throwapi.global.exception.exception;
 
 import org.springframework.http.HttpStatus;
 import sosteam.throwapi.global.exception.exception.CommonException;
 
 public class NoContentException extends CommonException {
     public NoContentException() {
-        super("", "", HttpStatus.NO_CONTENT);
+        super("NO_CONTENT","Nothing to show", HttpStatus.NO_CONTENT);
     }
 }
